+++ src/main/java/com/jmc/mazebank/Models/SavingsAccount.java
@@ -7,7 +7,7 @@ public class SavingsAccount extends Account{
     // The withdrawal limit from the savings
     private final DoubleProperty withdrawalLimit;
 
-    public SavingsAccount(String owner, String accountNumber, double balance, double withdrawalLimit) {
+    public SavingsAccount(String owner, String accountNumber, float balance, double withdrawalLimit) {
         super(owner, accountNumber, balance);
         this.withdrawalLimit = new SimpleDoubleProperty(this, "Withdrawal Limit", withdrawalLimit);
     }
